Fix NRE on BackupPage.
I didn't have this check in because I didn't think of this possibility.
If you have a program selected then add/remove it to the other list, the
first list is "active" but has no selection. If you then click Edit
Selected Program, an NRE is produced. I added null checks to fix this.