Start using mock tests.
Didn't have to make the API non-static like I thought I would,
thankfully.

One downside to the mock testing is that the stubbed methods have to be
virtual or in an interface. Right now I'm making them virtual, but maybe
later it would be better to have the database class implement an
interface.