Add ErrorMessage class.
This is meant to replace ErrorMessageBox, since it's more user friendly
to display an error inside the same window than create another dialog
that they have to close. I'll phase out and eventually delete
ErrorMessageBox soon.

The idea behind ErrorMessage is to create a standard look for an error
message. I plan on extending this as needed...it could include header
text (optionally, anyway), it should be closeable by the user, etc.

I need a way of easily keeping track of ErrorMessage instances in a
window so that the window doesn't create multiples of them, and so it
can destroy it when necessary.